backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
frozenlist==1.8.0
gotrue==2.12.4
greenlet==3.4.0
groq==0.9.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
langchain==0.2.12
langchain-community==0.2.10
langchain-core==0.2.30
langchain-groq==0.1.5
langchain-text-splitters==0.2.2
langsmith==0.1.147
lxml==5.2.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.26.4
orjson==3.11.7
packaging==24.2
postgrest==0.16.11
propcache==0.4.1
pydantic==2.10.4
pydantic-settings==2.6.1
pydantic_core==2.27.2
Pygments==2.20.0
PyJWT==2.12.1
PyMuPDF==1.24.3
PyMuPDFb==1.24.3
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.3
rank-bm25==0.2.2
realtime==1.0.6
requests==2.32.5
requests-toolbelt==1.0.0
rich==15.0.0
rich-toolkit==0.19.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==0.37.2
storage3==0.7.7
StrEnum==0.4.15
structlog==24.2.0
supabase==2.5.1
supafunc==0.4.7
tenacity==8.3.0
typer==0.24.1
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.29.0
watchfiles==1.1.1
websockets==12.0
yarl==1.23.0
